+++ lib/src/extensions.dart
@@ -1,10 +1,15 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import './extensions/utils/inkwell_manager.dart';
 
 part './extensions/build_context.dart';
+part './extensions/color.dart';
+part './extensions/edge_insets.dart';
 part './extensions/image_widget.dart';
 part './extensions/stateless_widget.dart';
+part './extensions/string.dart';
 part './extensions/text_style.dart';
 part './extensions/text_widget.dart';
 part './extensions/widget.dart';
+part './extensions/widget_list.dart';